+++ app/src/main/java/dev/tuandoan/expensetracker/di/DatabaseModule.kt
@@ -20,6 +20,7 @@ import dev.tuandoan.expensetracker.data.database.migration.MIGRATION_3_4
 import dev.tuandoan.expensetracker.data.database.migration.MIGRATION_4_5
 import dev.tuandoan.expensetracker.data.database.migration.MIGRATION_5_6
 import dev.tuandoan.expensetracker.data.database.migration.MIGRATION_6_7
+import dev.tuandoan.expensetracker.data.database.migration.MIGRATION_7_8
 import javax.inject.Singleton
 
 @Module
@@ -35,8 +36,15 @@ object DatabaseModule {
                 context = context,
                 klass = AppDatabase::class.java,
                 name = "expense_tracker_database",
-            ).addMigrations(MIGRATION_1_2, MIGRATION_2_3, MIGRATION_3_4, MIGRATION_4_5, MIGRATION_5_6, MIGRATION_6_7)
-            .build()
+            ).addMigrations(
+                MIGRATION_1_2,
+                MIGRATION_2_3,
+                MIGRATION_3_4,
+                MIGRATION_4_5,
+                MIGRATION_5_6,
+                MIGRATION_6_7,
+                MIGRATION_7_8,
+            ).build()
 
     @Provides
     fun provideTransactionDao(database: AppDatabase): TransactionDao = database.transactionDao()